Make alphabet handling more idiomatic and maintainable
And with luck, also more performant.

Alphabets are now represented as maps of indexed chars which we can more
easily slice, shuffle, reverse and look up.

Additionally, the rewrite ended up fixing the encoding of [40, 90], an
earlier example input which was failing, suggesting we're getting closer
to a correct implementation.