Make Sqids easier to use through macros
Allow for library consumers to `Use Sqids` and get a generated module to
be placed under their supervision tree, which takes care of storing and
loading state under the hood. `Sqids.encode(sqids, numbers)` becomes
`ConsumerSqids.encode(numbers)`, and so on.

Instances of Sqids on different modules will be completely independent
of each other.

Additionally, have test coverage of generated code by parameterizing the
cases to call both the direct and generated APIs.